stellar-data-pipeline/requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
PyYAML>=6.0.1
pytest>=7.4.0
pytest-cov>=4.1.0
